refactor integration handlers
